move options definition to its own file
